+++ spec/checkers/unique_index_checker_spec.rb
@@ -157,6 +157,22 @@ RSpec.describe DatabaseConsistency::Checkers::UniqueIndexChecker, :sqlite, :mysq
     end
   end
 
+  context 'when unique partial index exists', :postgresql do
+    before do
+      define_database_with_entity do |table|
+        table.integer :account_id
+        table.boolean :is_default
+        table.index %i[account_id], unique: true, name: index_name, where: 'is_default = true'
+      end
+    end
+
+    let(:klass) { define_class }
+
+    specify do
+      expect(checker.report).to be_nil
+    end
+  end
+
   context 'three attributes index' do
     before do
       define_database_with_entity do |table|