+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [Unreleased]
 
+- Add `ViewPrimaryKeyChecker` to enforce setting `primary_key` for models pointing to database views.
 - Fix `UniqueIndexChecker` to skip partial indexes (indexes with a `WHERE` clause).
 - Fix `MissingUniqueIndexChecker` to skip uniqueness validators with a `conditions` option.